feat(SKILL-136): close review-run completeness gaps
Persist per-lane review-run attribution (review_run_lanes) sourced from the
composed launch plan, carry lane columns on findings for pack/area
effectiveness, and always record terminal review state (review_finished_at,
non-null execution_mode) including zero-finding and telemetry-disabled runs.
Lane import replaces rather than appends. Schema ships as a numbered
migration plus unconditional column ensures.
