+++ runtime-kotlin/runtime-ports/src/main/kotlin/skillbill/ports/persistence/ReviewRepository.kt
@@ -8,36 +8,16 @@ import skillbill.review.model.FeedbackTelemetryOptions
 import skillbill.review.model.ImportedReview
 import skillbill.review.model.NumberedFinding
 import skillbill.review.model.ReviewFinishedTelemetry
-import skillbill.review.model.ReviewLaneEffectivenessRow
-import skillbill.review.model.ReviewRunLane
 
-interface ReviewRepository : WorkflowStatsRepository {
+interface ReviewRepository :
+  WorkflowStatsRepository,
+  ReviewRunCompletenessRepository {
   /** Stores only the schema-bounded accounting projection; content-bearing review objects cannot cross this seam. */
   fun saveAccounting(record: ReviewAccountingRecord) = Unit
 
   fun loadAccounting(reviewId: String): ReviewAccountingRecord? = null
 
   fun saveImportedReview(review: ImportedReview, sourcePath: String?)
-
-  /** Converges a run's per-lane attribution on the given plan-sourced set; safe to re-apply. */
-  fun replaceReviewRunLanes(runId: String, lanes: List<ReviewRunLane>) = Unit
-
-  fun fetchReviewRunLanes(runId: String): List<ReviewRunLane> = emptyList()
-
-  /**
-   * Records finding-to-lane attribution from the runtime's own merge result, keyed by finding id.
-   * Ingestion prefers it over provenance parsed out of review text.
-   */
-  fun recordFindingLaneAttribution(runId: String, attribution: Map<String, String>) = Unit
-
-  /** Pack-and-area effectiveness, grouped by canonical routed skill plus lane pack slug and area. */
-  fun reviewLaneEffectiveness(runId: String?): List<ReviewLaneEffectivenessRow> = emptyList()
-
-  /**
-   * Records the durable terminal state of a run — its finish timestamp and execution mode — for
-   * every run, including one that produced no findings and one imported with telemetry disabled.
-   */
-  fun ensureTerminalReviewState(runId: String, executionMode: String?) = Unit
 
   fun markOrchestrated(runId: String)
